# api-docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_store_asset_api_admin_store_assets__asset_kind__post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_store_asset_api_admin_store_assets__asset_kind__post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoreResponse": {
        "properties": {
          "favicon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favicon Url"
          },
          "footer_text": {
            "title": "Footer Text",
            "type": "string"
          },
          "gtm_container_id": {
            "title": "Gtm Container Id",
            "type": "string"
          },
          "home_cta_label": {
            "title": "Home Cta Label",
            "type": "string"
          },
          "home_description": {
            "title": "Home Description",
            "type": "string"
          },
          "home_eyebrow": {
            "title": "Home Eyebrow",
            "type": "string"
          },
          "home_headline": {
            "title": "Home Headline",
            "type": "string"
          },
          "homepage_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage Image Url"
          },
          "legal_name": {
            "title": "Legal Name",
            "type": "string"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "meta_pixel_id": {
            "title": "Meta Pixel Id",
            "type": "string"
          },
          "primary_color": {
            "title": "Primary Color",
            "type": "string"
          },
          "secondary_description": {
            "title": "Secondary Description",
            "type": "string"
          },
          "secondary_headline": {
            "title": "Secondary Headline",
            "type": "string"
          },
          "seo_description": {
            "title": "Seo Description",
            "type": "string"
          },
          "seo_title": {
            "title": "Seo Title",
            "type": "string"
          },
          "store_name": {
            "title": "Store Name",
            "type": "string"
          },
          "support_email": {
            "title": "Support Email",
            "type": "string"
          },
          "trust_items": {
            "items": {
              "type": "string"
            },
            "title": "Trust Items",
            "type": "array"
          },
          "whatsapp_message": {
            "title": "Whatsapp Message",
            "type": "string"
          },
          "whatsapp_number": {
            "title": "Whatsapp Number",
            "type": "string"
          }
        },
        "required": [
          "store_name",
          "legal_name",
          "primary_color",
          "logo_url",
          "favicon_url",
          "homepage_image_url",
          "whatsapp_number",
          "whatsapp_message",
          "support_email",
          "home_eyebrow",
          "home_headline",
          "home_description",
          "home_cta_label",
          "trust_items",
          "secondary_headline",
          "secondary_description",
          "footer_text",
          "seo_title",
          "seo_description",
          "gtm_container_id",
          "meta_pixel_id"
        ],
        "title": "StoreResponse",
        "type": "object"
      },
      "StoreUpdateRequest": {
        "properties": {
          "footer_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Footer Text"
          },
          "gtm_container_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gtm Container Id"
          },
          "home_cta_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Cta Label"
          },
          "home_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Description"
          },
          "home_eyebrow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Eyebrow"
          },
          "home_headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Headline"
          },
          "legal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Name"
          },
          "meta_pixel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Pixel Id"
          },
          "primary_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Color"
          },
          "secondary_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secondary Description"
          },
          "secondary_headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secondary Headline"
          },
          "seo_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Description"
          },
          "seo_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Title"
          },
          "store_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Name"
          },
          "support_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Email"
          },
          "trust_items": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Items"
          },
          "whatsapp_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Message"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          }
        },
        "title": "StoreUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "The same session token supplied as `Authorization: Bearer <token>`, for clients that cannot hold cookies. Accepted everywhere the cookie is.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Session cookie issued by `POST /api/auth/login`. `HttpOnly`, `SameSite=Strict`, and `Secure` outside local development, so a browser sends it automatically and page scripts cannot read it. This is the transport browser clients should use.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Cash-on-delivery commerce API for a single merchant: public landing delivery and checkout, plus the authenticated Administrator surface.",
    "title": "Easy Funnel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/store": {
      "get": {
        "description": "Return the same storefront configuration the public endpoint serves.\n\nSeparate from the public route so the editing surface reads through an\nauthenticated path and its access is audited like every other admin read.",
        "operationId": "admin_store_api_admin_store_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read storefront settings (admin)",
        "tags": [
          "store"
        ]
      },
      "patch": {
        "description": "Update storefront settings and return the stored result.\n\nA partial update: only the fields present in the body are written, so a\nclient can send one field without restating the rest. Values are validated\n(colors, WhatsApp number format, e-mail, tracking identifiers, lengths) and a\nrejection names the offending field.\n\nThese values seed from `STORE_*` environment variables **once**, on a fresh\ndatabase. After that this endpoint is the source of truth and a restart never\nreverts what the merchant saved here.",
        "operationId": "update_store_api_admin_store_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update storefront settings",
        "tags": [
          "store"
        ]
      }
    },
    "/api/admin/store/assets/{asset_kind}": {
      "post": {
        "description": "Replace one brand image and return the updated settings.\n\n`asset_kind` is one of `logo`, `favicon` or `homepage_image`. The upload is\nvalidated as an image, stored under a random opaque key, and the settings\nrecord is pointed at the new public URL. Send it as `multipart/form-data`\nwith the image in a `file` part.",
        "operationId": "upload_store_asset_api_admin_store_assets__asset_kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_kind",
            "required": true,
            "schema": {
              "title": "Asset Kind",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_store_asset_api_admin_store_assets__asset_kind__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Upload a brand image",
        "tags": [
          "store"
        ]
      }
    },
...
  },
...
}